+++ doc/version_history.rst
@@ -10,6 +10,7 @@ v1.20.0
 -------
 
 * Add the `rules.MTMountAzimuth` and `rules.MTM1M3Temperature` rules.
+* Improve the `rules.MTForceError` rule.
 
 v1.19.2
 -------